Added Search Container
